wip: oauth section refactoring
