--- src/scripts/modules/configurations/utils/authoriaztionConstants.js
@@ -1,5 +1,0 @@
-const authorizationConstants = {
-  OAUTH: 'oauth'
-};
-
-export default authorizationConstants;
